# 9_swag/1_api_swag/docs/swagger.yaml
basePath: /
definitions:
  dbase.CommentInfo:
    properties:
      body:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      postId:
        type: integer
    type: object
  dbase.PostInfo:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample of REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger REST API
  version: "1.0"
paths:
  /comment:
    post:
      consumes:
      - application/json
      description: create comment
      parameters:
      - description: post info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dbase.CommentInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Create comment
      tags:
      - comment
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Remove comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: get comment by ID
      operationId: get-comment-by-string
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Show comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: post info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dbase.CommentInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Update comment
      tags:
      - comment
  /post:
    get:
      consumes:
      - application/json
      description: get posts and comments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Show all posts and comments
      tags:
      - post
    post:
      consumes:
      - application/json
      description: create post
      parameters:
      - description: post info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dbase.PostInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Create post
      tags:
      - post
  /post/{id}:
    delete:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Remove post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: get post by ID
      operationId: get-post-by-int
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Show post
      tags:
      - post
    put:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: post info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dbase.PostInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Update post
      tags:
      - post
swagger: "2.0"